+++ sample/index.php
@@ -413,6 +413,33 @@ if (PHP_SAPI == 'cli') {
                             </div>
                         </div>
                     </li>
+                    <li class="list-group-item">
+                        <div class="row">
+                            <div class="col-md-8"><h5>Get only address balance</h5></div>
+                            <div class="col-md-4">
+                                <a href="addresses/GetOnlyBalance.php" class="btn btn-primary pull-left execute"> Try It
+                                    <i
+                                        class="fa fa-play-circle-o"></i></a>
+                                <a href="doc/addresses/GetOnlyBalance.html" class="btn btn-default pull-right">Source
+                                    <i
+                                        class="fa fa-file-code-o"></i></a>
+                            </div>
+                        </div>
+                    </li>
+                    <li class="list-group-item">
+                        <div class="row">
+                            <div class="col-md-8"><h5>Get multiple addresses balance at once (only balance)</h5></div>
+                            <div class="col-md-4">
+                                <a href="addresses/GetMultipleAddressesBalance.php"
+                                   class="btn btn-primary pull-left execute"> Try It <i
+                                        class="fa fa-play-circle-o"></i></a>
+                                <a href="doc/addresses/GetMultipleAddressesBalance.html"
+                                   class="btn btn-default pull-right">Source
+                                    <i
+                                        class="fa fa-file-code-o"></i></a>
+                            </div>
+                        </div>
+                    </li>
                 </ul>
             </div>